+++ Craftsman/Builders/Endpoints/ControllerBuilder.cs
@@ -29,6 +29,7 @@
             var wrapperClassPath = ClassPathHelper.WrappersClassPath(srcDirectory, "", projectBaseName);
             var featureClassPath = ClassPathHelper.FeaturesClassPath(srcDirectory, "", entityPlural, projectBaseName);
             var permissionsClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);
+            var rolesClassPath = ClassPathHelper.SharedKernelDomainClassPath(solutionDirectory, "");
             var permissionsUsing = usesJwtAuth 
                 ? $"{Environment.NewLine}using {permissionsClassPath.ClassNamespace};"
                 : string.Empty;
@@ -37,6 +38,7 @@
 
 using {featureClassPath.ClassNamespace};
 using {dtoClassPath.ClassNamespace};
+using {rolesClassPath.ClassNamespace};
 using {wrapperClassPath.ClassNamespace};{permissionsUsing}
 using System.Text.Json;
 using Microsoft.AspNetCore.JsonPatch;