+++ Craftsman/Builders/Tests/FunctionalTests/PatchEntityTestBuilder.cs
@@ -26,8 +26,10 @@
             var fakerClassPath = ClassPathHelper.TestFakesClassPath(testDirectory, "", entity.Name, projectBaseName);
             var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", entity.Name, projectBaseName);
             var permissionsClassPath = ClassPathHelper.PolicyDomainClassPath(testDirectory, "", projectBaseName);
+            var rolesClassPath = ClassPathHelper.SharedKernelDomainClassPath(solutionDirectory, "");
+            
             var permissionsUsing = isProtected 
-                ? $"{Environment.NewLine}using {permissionsClassPath.ClassNamespace};"
+                ? $"{Environment.NewLine}using {permissionsClassPath.ClassNamespace};{Environment.NewLine}using {rolesClassPath.ClassNamespace};"
                 : string.Empty;
 
             var authOnlyTests = isProtected ? $@"