+++ Craftsman/Builders/Tests/IntegrationTests/UserPolicyHandlerIntegrationTests.cs
@@ -27,6 +27,7 @@
             var policyDomainClassPath = ClassPathHelper.PolicyDomainClassPath(srcDirectory, "", projectBaseName);
             var entityClassPath = ClassPathHelper.EntityClassPath(testDirectory, "", "RolePermissions", projectBaseName);
             var dtoClassPath = ClassPathHelper.DtoClassPath(solutionDirectory, "", "RolePermission", projectBaseName);
+            var rolesClassPath = ClassPathHelper.SharedKernelDomainClassPath(solutionDirectory, "");
 
             return @$"namespace {classPath.ClassNamespace};
 
@@ -34,6 +35,7 @@ using {servicesClassPath.ClassNamespace};
 using {policyDomainClassPath.ClassNamespace};
 using {entityClassPath.ClassNamespace};
 using {dtoClassPath.ClassNamespace};
+using {rolesClassPath.ClassNamespace};
 using Bogus;
 using FluentAssertions;
 using HeimGuard;